Set up Dependabot version updates
## Problem Statement

We're getting ready to pin our `package.json`'s dependencies to specific version. We need a way to be automatically notified when those dependencies have new updates. Let's try using dependabot for that as well.

## Proposed Changes

Add a `dependabot.yml` which will configure it to check on our dependencies weekly for updates.

https://docs.github.com/en/code-security/dependabot/dependabot-version-updates/about-dependabot-version-updates

## Verification Steps

- [ ] Not sure if this can be verified before merge, but maybe you can think of something? Otherwise let's give it a try!